+++ addon/components/item-picker/component.js
@@ -48,6 +48,15 @@ export default Ember.Component.extend({
     }
   }),
 
+  previewUrl: Ember.computed('portalHostName', 'currentItem.id', 'currentItem.url', function () {
+    const url = this.get('portalHostName');
+    if (url) {
+      return `${url}/home/item.html?id=${this.get('currentItem.id')}`;
+    } else {
+      return this.get('currentItem.url');
+    }
+  }),
+
   inputElementId: Ember.computed(function () {
     return `${this.get('elementId')}-search-items`;
   }),